misc: vexpress/spc: attempt to init the spc driver from vexpress_spc_check_loaded()
Even if it is initialized via early_INITCALL(), there are other pieces
of code that also have to be initialized with early_initcall() and
the relative call ordering may not always be satisfied.

So try to initialize the SPC code when vexpress_spc_check_loaded() is
called if that has not been done yet.

Signed-off-by: Nicolas Pitre <nico@linaro.org>