ARM: bL_entry: Fix incorrect coverage with initialising bL_sync struct
The indexing is incorrect when initialising the bL_sync structure
in bL_cluster_sync_init().

For debugging purposes, 0 is not a valid value for any of these
fields and is therefore not a sensible default.

This patch fixes the indexing to iterate over all the fields and
initialise them to valid values.

Signed-off-by: Dave Martin <dave.martin@linaro.org>