ARM: bL_platsmp.c: make sure the GIC interface of a dying CPU is disabled
Otherwise there might be some interrupts or IPIs becoming pending and the
CPU will not enter low power mode when doing a WFI.  The effect of this
is a CPU that loops back into the kernel, go through the first man
election, signals itself as alive,  and prevent the cluster from being
shut down.

This is hacky and needs a better solution.