Revert "ARM: common: add GIC bybass disable on GIC CPU IF save function"
On some platforms the GIC CPU IF must be disabled on power down to make
sure wfi state is not exited on others the GIC CPU IF must remain
enabled to ensure correct propagation of IRQs to the corresponding CPU
input pin.

This patch removes common code in GIC PM notifier that disables the
GIC CPU IF on power down.

Code is already present in the GIC driver that allows to disable GIC
CPU IF on a per-platform basis.

Signed-off-by: Lorenzo Pieralisi <lorenzo.pieralisi@arm.com>